move some requests to utils
